+++ QiniuSDK/Utils/QNPHAssetFile.m
@@ -82,7 +82,7 @@
     return self;
 }
 
-- (NSData *)read:(long)offset
+- (NSData *)read:(long long)offset
             size:(long)size
            error:(NSError **)error {
     
@@ -90,8 +90,8 @@
     @try {
         [_lock lock];
         if (_assetData != nil && offset < _assetData.length) {
-            NSInteger realSize = MIN(size, _assetData.length - offset);
-            data = [_assetData subdataWithRange:NSMakeRange(offset, realSize)];
+            NSUInteger realSize = MIN((NSUInteger)size, _assetData.length - (NSUInteger)offset);
+            data = [_assetData subdataWithRange:NSMakeRange((NSUInteger)offset, realSize)];
         } else if (_file != nil && offset < _fileSize) {
             [_file seekToFileOffset:offset];
             data = [_file readDataOfLength:size];
